requirements.txt
numpy~=1.21.6
pandas==1.3.5
scipy==1.7.3
tensorflow-gpu==2.3.4
tqdm~=4.64.0

dgl==1.0.3
torch==1.13.0
pyyaml==6.0